Split KLMS auth god module into SSO, browser, and OTP parts
Extract Playwright launch/install into auth_browser, HTML/SSO and
easy-login flows into auth_sso, and email-OTP worker IPC into auth_otp,
leaving AuthService as a composing facade with stable re-exports.
